% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{eve<-}
\alias{eve<-}
\title{Set eve of particles}
\usage{
eve(object) <- value
}
\arguments{
\item{object}{Object to set eve for.}

\item{value}{Value of new eve.}
}
\description{
Set eve of particles
}
