% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{eve.particles}
\alias{eve.particles}
\title{Eve of particles}
\usage{
\method{eve}{particles}(object)
}
\arguments{
\item{object}{Object to return eve for.}
}
\value{
eve, original ancestor of particles.
}
\description{
Eve of particles
}
