% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve-variance.R
\name{eve_var_est}
\alias{eve_var_est}
\title{Variance estimator using ancestory}
\usage{
eve_var_est(particles, log_z, num_iter)
}
\arguments{
\item{particles}{Particles to calculate variance estimate for.}

\item{log_z}{log of normalising constant estimated by algorithm.}

\item{num_iter}{Number of SMC iterations.}
}
\value{
estimate of MC variance of normalising constant.
}
\description{
See: Supplementary material for 'Variance estimation in the particle filter'
}
