% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra.R
\name{generator_matrix_lotka_volterra_ctmc}
\alias{generator_matrix_lotka_volterra_ctmc}
\title{Generator matrix for Lotka-Volterra continuous-time Markov chain}
\usage{
generator_matrix_lotka_volterra_ctmc(theta, y1_min = 0L, y1_max,
  y2_min = 0L, y2_max)
}
\arguments{
\item{theta}{Numeric vector of length 3.}

\item{y1_min}{Min for y1, default is 0.}

\item{y1_max}{Max for y1.}

\item{y2_min}{Min for y2, default is 0.}

\item{y2_max}{Max for y2.}
}
\value{
Sparse matrix
}
\description{
Generator matrix for Lotka-Volterra continuous-time Markov chain
}
