% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-memoised-log-likelihood-interface.R
\name{log_approx_likelihood_anneal_func_da}
\alias{log_approx_likelihood_anneal_func_da}
\title{Generic interface to annealed log-likeihood functions with memoisation: Approximate likelihood annealing.}
\usage{
log_approx_likelihood_anneal_func_da(
  log_likelihood,
  log_like_approx,
  log_prior,
  max_approx_anneal
)
}
\arguments{
\item{log_likelihood}{(Full) log-likelihood function.}

\item{log_like_approx}{Approximate log-likelihood function.}

\item{log_prior}{Log-prior function.}

\item{max_approx_anneal}{Maximum value of temperature during approximate anneal.}
}
\value{
Value of log-likelihood function.
}
\description{
NOT IN USE: Delete after confirmation.
}
\details{
Approximate posterior -> Full posterior
}
