% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra-ctmc.R
\name{log_lhood_lotka_volterra_ctmc}
\alias{log_lhood_lotka_volterra_ctmc}
\title{Log-likelihood for Lotka-Volterra continuous-time Markov chain}
\usage{
log_lhood_lotka_volterra_ctmc(theta, y1, y2, times, y1_max, y2_max, ...)
}
\arguments{
\item{theta}{Parameters, numeric vector of length 3.}

\item{y1}{Observed y1.}

\item{y2}{Observed y2.}

\item{times}{Observed times.}

\item{y1_max}{Max for y1, integer.}

\item{y2_max}{Max for y2, integer.}

\item{...}{To be passed to prob function.}
}
\value{
log-likelihood (numeric).
}
\description{
Log-likelihood for Lotka-Volterra continuous-time Markov chain
}
