% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra-lna.R
\name{log_lhood_lotka_volterra_lna}
\alias{log_lhood_lotka_volterra_lna}
\title{Log-likelihood for Lotka-Volterra Linear Noise Approximation of CTMC}
\usage{
log_lhood_lotka_volterra_lna(theta, y1, y2, times)
}
\arguments{
\item{theta}{Parameters, numeric vector of length 3.}

\item{y1}{Observed y1.}

\item{y2}{Observed y2.}

\item{times}{Observed times.}
}
\value{
log-likelihood (numeric)
}
\description{
Log-likelihood for Lotka-Volterra Linear Noise Approximation of CTMC
}
