% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-conservative-DAMH.R
\name{mh_da_step_bglr}
\alias{mh_da_step_bglr}
\title{Delayed-acceptance MH for symmetric kernel.}
\usage{
mh_da_step_bglr(
  new_particles,
  old_particles,
  var,
  temp,
  loglike,
  pre_trans = identity,
  time_on = T,
  approx_ll_weights = NULL,
  mvn_step_scale = NULL,
  reg_warnings = T
)
}
\arguments{
\item{new_particles}{Locations for proposed new particles.}

\item{old_particles}{Locations for current/old particles.}

\item{var}{Empirical variance for Mahalanobis distance calculation.}

\item{temp}{Current SMC temperature.}

\item{loglike}{Generic log-likelihood interface. Vectorised!}

\item{pre_trans}{Pre-transformation for approximate log-likelihood.}

\item{time_on}{Logical. Should computation time be recorded?}

\item{approx_ll_weights}{Weights for approximate likelihood.}

\item{mvn_step_scale}{Step scale used for to generate each proposed particle.}

\item{reg_warnings}{Print regression warnings.}
}
\value{
List.
}
\description{
Uses http://aimsciences.org//article/doi/10.3934/fods.2019005
}
\details{
Assumes symmetric or independent kernel.
}
