% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-standard-MH.R
\name{mh_step}
\alias{mh_step}
\title{MH for symmetric kernel.}
\usage{
mh_step(new_particles, old_particles, var, temp, loglike, type,
  time_on = T)
}
\arguments{
\item{new_particles}{Locations for proposed new particles.}

\item{old_particles}{Locations for current/old particles.}

\item{var}{Empirical variance for Mahalanobis distance calculation.}

\item{temp}{Current SMC temperature.}

\item{loglike}{Generic log-likelihood interface.}

\item{type}{Which log-likelihood to use. approx or full}

\item{time_on}{Logical. Should computation time be recorded?}
}
\value{
List.
}
\description{
Assumes symmetric or independent kernel.
}
