% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-select-best-tuning-params.R
\name{min_da_mh_cost}
\alias{min_da_mh_cost}
\title{Find tuning parameter with minimum computation cost under DAMH,}
\usage{
min_da_mh_cost(min_T, surrogate_acceptance, surrogate_cost, full_cost)
}
\arguments{
\item{min_T}{Minimum time taken to perform MH iteration for each parameter.}

\item{surrogate_acceptance}{Average surrogate acceptance rate for each tuning parameter.}

\item{surrogate_cost}{Average surrogate acceptance rate for each tuning parameter.}

\item{full_cost}{Average cost of using surrogate.}
}
\value{
Index for which parameter to use.
}
\description{
Find tuning parameter with minimum computation cost under DAMH,
}
