% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-select-best-tuning-params.R
\name{min_mh_cost}
\alias{min_mh_cost}
\title{Find tuning parameter with minimum computation cost under MH}
\usage{
min_mh_cost(min_T)
}
\arguments{
\item{min_T}{Minimum time taken to perform MH iteration for each parameter.}
}
\value{
Index for which parameter to use.
}
\description{
Find tuning parameter with minimum computation cost under MH
}
