% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn-jitter.R
\name{mvn_jitter}
\alias{mvn_jitter}
\title{Jitter particles object with MVN kernel.}
\usage{
mvn_jitter(particles, step_scale = 1, var = NULL, prec = NULL,
  var_chol = NULL, prec_chol = NULL)
}
\arguments{
\item{particles}{Particles to be jittered.}

\item{step_scale}{How much to scale the standard deviation by.}

\item{var}{Variance matrix.}

\item{prec}{Precision matrix.}

\item{var_chol}{Cholesky decomposition of variance matrix (upper triangular).}

\item{prec_chol}{Cholesky decomposition of precision matrix (upper triangular).}
}
\description{
\code{mvn_jitter} returns particles that have been jittered.
}
