% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{names.particles}
\alias{names.particles}
\title{Extract names of elements in each particle.}
\usage{
\method{names}{particles}(x, ...)
}
\arguments{
\item{x}{Particle.}

\item{...}{Not in use.}
}
\description{
Extract names of elements in each particle.
}
