% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{num_particles}
\alias{num_particles}
\title{Number of particles per component.}
\usage{
num_particles(x)
}
\arguments{
\item{x}{particles object.}
}
\description{
Get the number of particles for a particular object.
}
