% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{p_components}
\alias{p_components}
\title{Particle components.}
\usage{
p_components(x, attr = NULL)
}
\arguments{
\item{x}{particles object.}

\item{attr}{the particular attribute of the component to access.}
}
\description{
Extract information about the components of a \code{particles} object.
}
