% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-dim.R
\name{p_dim}
\alias{p_dim}
\title{Get the dimensions of the components of particles.}
\usage{
p_dim(x, ...)
}
\arguments{
\item{x}{Object to check dimesion of.}

\item{...}{Extra arguments.}
}
\description{
Get the dimensions of the components of particles.
}
