% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-shape.R
\name{p_shape}
\alias{p_shape}
\title{Reshape components of particles to matrix.}
\usage{
p_shape(x, ...)
}
\arguments{
\item{x}{Object to shape.}

\item{...}{Extra arguments.}
}
\description{
Reshape components of particles to matrix.
}
