% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle-apply.R
\name{papply}
\alias{papply}
\title{Apply function to all particles}
\usage{
papply(particles, fun, ...)
}
\arguments{
\item{particles}{Particle object.}

\item{fun}{Function to apply.}

\item{...}{}
}
\value{
Function evaluated at each row.
}
\description{
Function must be formatted to take rows from particle matrix.
}
