% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{particles}
\alias{particles}
\title{Create particles object.}
\usage{
particles(...)
}
\arguments{
\item{...}{Names and types of object that is being used for each component in a particle.}
}
\description{
\code{particles} returns a particle with specified names.
}
\examples{
num_particles <- 10
len <- 2
rv <- matrix(rnorm(num_particles * len), nrow = num_particles, ncol = len)
prts <- particles(beta = rv)
}
