% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{print.particles}
\alias{print.particles}
\title{Print particles object.}
\usage{
\method{print}{particles}(x, ...)
}
\arguments{
\item{x}{particles to be printed.}

\item{...}{Arguments to be passed to \code{str}.}
}
\description{
\code{particles} returns a particle with specified names.
}
\examples{
x <- particles(beta = matrix(rnorm(20), nrow = 10, ncol = 2))
print(x)
}
