% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{replace_particles}
\alias{replace_particles}
\title{Replace (subset of) old set of particles with (subset of) new.}
\usage{
replace_particles(new_particles, old_particles, index, reweight = T)
}
\arguments{
\item{new_particles}{New particle object}

\item{old_particles}{Old particle object}

\item{index}{Indices to use.}

\item{reweight}{Should new particles be assinged a uniform weight?}
}
\value{
Particles object.
}
\description{
Replaces the index of old particles with index of new particles. Weights do not update.
}
