% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_multinomial}
\alias{resample_multinomial}
\title{Generic multinomial resample}
\usage{
resample_multinomial(x)
}
\arguments{
\item{x}{Object to resample elements of.}
}
\value{
Resampled object
}
\description{
Generic multinomial resample
}
