% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample_stratified}
\alias{resample_stratified}
\title{Generic stratified resample}
\usage{
resample_stratified(x, num_strata)
}
\arguments{
\item{x}{Object to resample elements of.}

\item{num_strata}{Number of strata.}
}
\value{
Resampled object
}
\description{
Generic stratified resample
}
