% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-cluster-cov.R
\name{robust_mean_cov}
\alias{robust_mean_cov}
\title{Robust covariance and correlation estimation from matrix}
\usage{
robust_mean_cov(X, ncluster, ...)
}
\arguments{
\item{X}{A matrix, rows are observations.}

\item{ncluster}{Number of clusters to use. If missing \code{ClusterR::Optimal_Clusters_KMeans} is used.}

\item{...}{Settings for \code{stats::cov.wt}.}
}
\value{
Covariance or correlation matrix.
}
\description{
Uses a clustering algorithm to estimate means, then estimate the covariance or correlation matrix.
Better for multimodal distributions.
}
