% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-run-smc-da.R
\name{run_smc_da}
\alias{run_smc_da}
\title{Run SMC sampler with DA (or possibly not DA methods)}
\usage{
run_smc_da(
  num_p,
  step_scale_set,
  use_da,
  use_approx = F,
  start_from_approx = F,
  refresh_ejd_threshold,
  par_start,
  log_prior,
  log_like,
  log_like_approx,
  log_like_approx_comps = NULL,
  draw_prior,
  start_from_approx_fit,
  optimise_pre_approx_llhood_transformation,
  optimise_approx_llhood_weights,
  find_best_step_scale,
  max_anneal_temp = 1,
  save_post_interface,
  cores = 1L,
  verbose = F,
  ...
)
}
\arguments{
\item{num_p}{Number of particles.}

\item{step_scale_set}{Set  of scales used in MVN kernel.}

\item{use_da}{Logical. Use Delayed-acceptance?}

\item{use_approx}{Logical. Use approximate likelihood?}

\item{start_from_approx}{Logical. Start from approximate likelihood the transition to full with delayed-acceptance?}

\item{refresh_ejd_threshold}{Distance required to travel for expected jumping distance.}

\item{par_start}{Approximate likelihood optimisation parameter starting value.}

\item{log_prior}{Function for log-prior.}

\item{log_like}{Function for (full) log-likelihood.}

\item{log_like_approx}{Function for approximate log-likelihood.}

\item{log_like_approx_comps}{Function for approximate log-likelihood that returns components.}

\item{draw_prior}{Function to draw random sample from prior.}

\item{start_from_approx_fit}{SMC object (list) with approximate fit if start_from_approx is TRUE.}

\item{optimise_pre_approx_llhood_transformation}{Function to optimise pre-transformation for approximate likelihood.}

\item{optimise_approx_llhood_weights}{Function to optimise weights of approximate log-likelihood.}

\item{find_best_step_scale}{Function to find best step scale.}

\item{max_anneal_temp}{Temperature to anneal likelihood (posterior) to. Max 1.}

\item{save_post_interface}{Logical. Should the memoised interface to the likelihood functions be returned (very large).}

\item{cores}{Number of cores.}

\item{verbose}{Logical. Should the SMC iterations update be printed to console?}

\item{...}{Arguments to be passed to functions.}
}
\value{
A list
}
\description{
Run SMC sampler with DA (or possibly not DA methods)
}
