% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{select_reweight_particles}
\alias{select_reweight_particles}
\title{Select new set of particles from old, reweight.}
\usage{
select_reweight_particles(object, index, reweight = T)
}
\arguments{
\item{object}{Particle object}

\item{index}{Indices to use.}

\item{reweight}{Should new particles be assinged a uniform weight?}
}
\value{
Particles object.
}
\description{
Select new set of particles from old, reweight.
}
