% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-ctmc.R
\name{simulate_ctmc}
\alias{simulate_ctmc}
\title{Simulate continuous-time Markov chain}
\usage{
simulate_ctmc(generator_matrix, initial_state_prob, initial_state,
  total_time)
}
\arguments{
\item{generator_matrix}{Square generator matrix (possibly sparse).}

\item{initial_state_prob}{Probability vector for intial state of process.}

\item{initial_state}{Initial state of process.}

\item{total_time}{Total time that CTMC can run for.}
}
\value{
Trajectory (integer vector)
}
\description{
Simulate continuous-time Markov chain
}
