% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra-ctmc.R
\name{simulate_lotka_volterra_ctmc}
\alias{simulate_lotka_volterra_ctmc}
\title{Simulate continuous-time Markov chain of Lotka-Volterra model}
\usage{
simulate_lotka_volterra_ctmc(
  theta,
  y1_min = 0,
  y2_min = 0,
  y1_max,
  y2_max,
  initial_y1,
  initial_y2,
  initial_state_prob,
  total_time
)
}
\arguments{
\item{theta}{Numeric vector of length 3.}

\item{y1_min}{Min for y1, integer, default is 0.}

\item{y2_min}{Min for y2, integer, default is 0.}

\item{y1_max}{Max for y1, integer.}

\item{y2_max}{Max for y2, integer.}

\item{initial_y1}{Initial state of process y1.}

\item{initial_y2}{Initial state of process y2.}

\item{initial_state_prob}{Probability matrix (y1, y2) for intial state of process.}

\item{total_time}{Total time that CTMC can run for.}
}
\value{
Trajectory (integer vector)
}
\description{
Simulate continuous-time Markov chain of Lotka-Volterra model
}
