% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcdar-package.R
\docType{package}
\name{smcdar}
\alias{smcdar}
\title{\code{smcdar}: A package for delayed-acceptance SMC.}
\description{
smcdar...
}
\section{\code{smcdar} objects}{

\itemize{
   \item \code{\link{particles}}
}
}

\section{\code{smcdar} functions}{

\itemize{
   \item \code{\link{mvn_jitter}}
   \item \code{\link{num_particles}}
   \item \code{\link{p_components}}
   \item \code{\link{p_dim}}
   \item \code{\link{p_shape}}
}
}

