% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-required-MH-steps.R
\name{time_steps_to_min_quantile_dist_bootstrap}
\alias{time_steps_to_min_quantile_dist_bootstrap}
\title{Calculate number of MH steps required until threshold is met.}
\usage{
time_steps_to_min_quantile_dist_bootstrap(dist, D, rho, max_T = 10,
  boot_samples = 100)
}
\arguments{
\item{dist}{(Expected) distances travelled by each particle.}

\item{D}{Threshold distance required to travel in total.}

\item{rho}{Probability level of exceeding threshold distance.}

\item{max_T}{Maximum number of iterations to compute before "failing".}

\item{boot_samples}{Number of bootstrap samples to use.}
}
\value{
List with results for each parameter.
}
\description{
Bootstrap version.
}
