% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA-required-MH-steps.R
\name{time_steps_to_min_quantile_dist_normal}
\alias{time_steps_to_min_quantile_dist_normal}
\title{Calculate number of MH steps required until threshold is met.}
\usage{
time_steps_to_min_quantile_dist_normal(dist, accept, D, rho, max_T = 10)
}
\arguments{
\item{dist}{Proposed distances travelled by each particle.}

\item{accept}{Whether a proposal was accepted or rejected.}

\item{D}{Threshold distance required to travel in total.}

\item{rho}{Probability level of exceeding trheshold distance.}

\item{max_T}{Maximum number of iterations to compute before "failing".}
}
\value{
List with results for each parameter.
}
\description{
Normal version.
}
