% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{weights<-}
\alias{weights<-}
\title{Set  weights}
\usage{
weights(object, ..., log) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Other parameters.}

\item{log}{Are log-weights being used?}

\item{value}{New weights.}
}
\description{
Set  weights
}
