% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{weights.particles}
\alias{weights.particles}
\title{Get particle weights}
\usage{
\method{weights}{particles}(object, ..., log = FALSE, normalise = TRUE)
}
\arguments{
\item{object}{Particles object.}

\item{...}{Other parameters.}

\item{log}{Return log weights?}

\item{normalise}{Normalise weights? Sum to one.}
}
\value{
Weights
}
\description{
Get particle weights
}
