% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka-volterra-lna.R
\name{d_dt_lotka_volterra_lna}
\alias{d_dt_lotka_volterra_lna}
\title{Time-derivative for LNA of Lotka-Volterra CTMC}
\usage{
d_dt_lotka_volterra_lna(t, y, theta)
}
\arguments{
\item{t}{time parameter (not used, but needed for ode solver).}

\item{y}{Population vector.}

\item{theta}{Parameter vector.}
}
\value{
Derivative vector for LNA approximation with components: (y1, y2, v11, v12, v22).
}
\description{
Time-derivative for LNA of Lotka-Volterra CTMC
}
