% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle-apply.R
\name{papply}
\alias{papply}
\title{Apply function to all particles}
\usage{
papply(particles, fun, comp_time = F, cores = 1L, ...)
}
\arguments{
\item{particles}{Particle object.}

\item{fun}{Function to apply.}

\item{comp_time}{Record time taken for comutation? Logical.}

\item{cores}{Use multicore to evaluate? Uses parallel package.}

\item{...}{Other arguments to pass to \code{fun}.}
}
\value{
Function evaluated at each row.
}
\description{
Function must be formatted to take rows from particle matrix.
}
