% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-cluster-cov.R
\name{robust_mean_cov}
\alias{robust_mean_cov}
\title{Robust covariance and correlation estimation from matrix}
\usage{
robust_mean_cov(X, max_ncluster = 5, ...)
}
\arguments{
\item{X}{A matrix, rows are observations.}

\item{max_ncluster}{Max number of clusters to consider.}

\item{...}{Settings for \code{stats::cov.wt}.}
}
\value{
Covariance or correlation matrix.
}
\description{
Uses a clustering algorithm to estimate means, then estimate the covariance or correlation matrix.
Better for multimodal distributions. Uses \code{fpc::kmeansruns} to determine number of clusters.
}
