% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-object.R
\name{weights<-.particles}
\alias{weights<-.particles}
\title{Set particle weights}
\usage{
\method{weights}{particles}(object, log = FALSE) <- value
}
\arguments{
\item{object}{Particles object.}

\item{log}{Input is log weights?}

\item{value}{New particle weights.}
}
\description{
Set particle weights
}
